% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ponyexpress-package.r
\docType{data}
\name{glitter_template}
\alias{glitter_template}
\title{HTML email template with some glitter}
\format{An object of class \code{character} of length 1.}
\usage{
glitter_template
}
\description{
HTML email template with some glitter
}
\keyword{datasets}
