% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/note_send.R
\name{note_send}
\alias{note_send}
\title{Sends email from notifier object}
\usage{
note_send(edat, type = "text", secret, errors = "errors.csv",
  sent = "sent-emails.rds")
}
\arguments{
\item{edat}{tibble outputted from note_create}

\item{secret}{path to secret.json file}

\item{errors}{Path for file to record error messages}

\item{sent}{Path for sent emails to be saved. You will need this later as a log of sent emails. Note that successfully sent emails will also be in your Gmail sent folder.}
}
\description{
Sends email from notifier object
}
