% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parcel_create.R
\name{parcel_create}
\alias{parcel_create}
\title{Create a parcel to email}
\usage{
parcel_create(df, sender_name = NULL, sender_email = NULL, subject = NULL,
  bcc = NULL, template = NULL)
}
\arguments{
\item{df}{data frame, information to be fed to the template (required)}

\item{sender_name}{character, the sender's name (required)}

\item{sender_email}{character, the sender's email (required)}

\item{subject}{character, the email's subject}

\item{bcc}{character, email address to bcc}

\item{template}{character, template for email, using glue syntax (see example)}
}
\value{
data frame with information to email
}
\description{
Create a parcel to email
}
\examples{
df <- data.frame(
name = c("Lucy", "Karthik"),
email = c("example@email.com", "example@email.com")
)

template <- "
Dear {name},

This is a friendly email from me.

XO,
Lucy"

parcel_create(df,
            sender_name = "Lucy",
            sender_email = "lucydagostino@gmail.com",
            subject = "Happy email!",
            template = template)

}
