% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parcel_preview.R
\name{parcel_preview}
\alias{parcel_preview}
\title{Parcel Preview}
\usage{
parcel_preview(email = NULL)
}
\arguments{
\item{email}{object of class `ponyexpress`, can be obtained from the
`parcel_create()` function.}
}
\description{
Preview the parcel in a new browser window.
}
