% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parcel_send.R
\name{parcel_send}
\alias{parcel_send}
\title{Sends email from notifier object}
\usage{
parcel_send(edat, type = "text", secret, errors = "errors.csv",
  sent = "sent-emails.rds")
}
\arguments{
\item{edat}{tibble outputted from note_create}

\item{type}{Type of email. Either text or html.}

\item{secret}{path to secret.json file}

\item{errors}{Path for file to record error messages}

\item{sent}{Path for sent emails to be saved. You will need this later as a log of sent emails.}
}
\description{
Sends email from notifier object
}
\examples{
# First let's generate a sample output from note_create

sample_email <-
data.frame(To = c("N. Bluth <lucille.bluth@gmail.com>", "G.O.B. <gob@imoscardotcom.com>"),
From = c("Karthik Ram <karthik.ram@gmail.com>"),
Subject = "This is the subject",
Body = "This is the message")
# Now send emails with
note_send(sample_email, secret)
}
